Update Makefile and ShellCommand.rb
The Makefile has been updated to include a new target called "test_large". This target runs a GitHub workflow called "test-dataset.yml" with the following parameters: commit_sha (set to the current git revision) and dataset_name (set to "coderbot_dataset_140"). After running the workflow, it waits for 1 second and then prints the URL of the first run using the "gh run list" command.

The ShellCommand.rb file has been updated to include additional information in the error message when a command fails. The error message now includes the STDOUT and STDERR of the failed command.

Commit message: "Update Makefile and ShellCommand.rb